Restructure to multiple files
