+++ binding.gyp
@@ -3,8 +3,10 @@
         {
             "target_name": "node_gtk",
             "sources": [
+                "src/async_call_environment.cc",
                 "src/boxed.cc",
                 "src/callback.cc",
+                "src/callback_wrapper.cc",
                 "src/closure.cc",
                 "src/debug.cc",
                 "src/error.cc",
