+++ src/gi.cc
@@ -2,6 +2,7 @@
 #include <node.h>
 #include <nan.h>
 
+#include "async_call_environment.h"
 #include "boxed.h"
 #include "closure.h"
 #include "debug.h"
@@ -338,7 +339,7 @@ NAN_METHOD(GetModuleCache) {
 }
 
 void InitModule(Local<Object> exports, Local<Value> module, void *priv) {
-    GNodeJS::Closure::Initialize();
+    GNodeJS::AsyncCallEnvironment::Initialize();
 
     NAN_EXPORT(exports, Bootstrap);
     NAN_EXPORT(exports, GetModuleCache);
